first implementation of play mode
added button to enter playmode

playmode prevents any objects from being dragged, created, rotated or deleted.

spawns a marble at the begining of the maze

you can't move parts of the level that you didn't create at any time

Fixed bugs where errors are thrown if you try to drag on any object in the scene that is not a block. As well as being able to zoom too far in.